requirements.txt
asgiref==3.6.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
Django==4.2.1
django-cors-headers==3.14.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
psycopg2-binary==2.9.6
pycparser==2.21
python-decouple==3.8
pytz==2023.3
PyYAML==6.0
requests==2.30.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.2
